requirements.txt
streamlit~=1.42.2
openai~=1.56.0
pandas~=2.2.3
numpy~=1.26.4
fpdf2~=2.8.1
pdfplumber~=0.11.4
pytesseract~=0.3.13
beautifulsoup4~=4.12.3
requests~=2.32.3
matplotlib~=3.9.4
Pillow~=10.4.0
google-search-results~=2.4.2
python-dotenv~=1.0.1
xlsxwriter~=3.2.0
gspread~=6.1.4
serpapi~=0.1.5
google-auth~=2.35.0
google-generativeai~=0.8.3
reportlab~=4.2.5